finished integration testing

+++ backend/src/main/java/com/rpissarra/recipe/RecipeService.java
@@ -14,7 +14,6 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Transactional
@@ -58,21 +57,21 @@ public class RecipeService {
         if (recipeRegistrationRequest.ingredients().isEmpty() || recipeRegistrationRequest.steps().isEmpty()) {
             throw new RequestValidationException("The list of ingredients/steps can't be empty");
         }
-        Recipe recipe = new Recipe();
         Date createDate = new Date();
-        recipe.setName(recipeRegistrationRequest.name().toLowerCase());
-        recipe.setCreatedate(createDate);
-        recipe.setUpdatedate(null);
+        Recipe recipe = new Recipe(
+                recipeRegistrationRequest.name().toLowerCase(),
+                createDate
+        );
         recipeRepository.save(recipe);
 
-        List<Ingredients> ingredientsList = convertToIngredientsList(recipeRegistrationRequest.ingredients(), recipe);
+        List<Ingredients> ingredientsList = convertToIngredientsList(recipeRegistrationRequest.ingredients(), recipe, createDate);
         ingredientRepository.saveAll(ingredientsList);
 
-        List<Steps> stepsList = convertToStepsList(recipeRegistrationRequest.steps(), recipe);
-        stepsRepository.saveAll(stepsList);
-    }
-    public Optional<Recipe> getRecipeById(Long id) {
-        return recipeRepository.findById(id);
+        stepsRepository.save(new Steps(
+                recipeRegistrationRequest.steps(),
+                createDate,
+                recipe
+        ));
     }
 
     public List<RecipeDTO> getAllRecipesByName(String name) {
@@ -99,11 +98,13 @@ public class RecipeService {
     }
 
     public void updateRecipe(Long id, RecipeUpdateRequest recipeUpdateRequest) {
-        Recipe recipe = getRecipeById(id)
-                .orElseThrow(
-                        () -> new ResourceNotFoundException("Could not find any Recipe with id [%s]".formatted(id))
-                );
-
+        Recipe recipe = null;
+        Date updateDate = new Date();
+        if (recipeRepository.existsRecipeByIdrecipe(id)) {
+            recipe = recipeRepository.findById(id).get();
+        } else {
+            throw new ResourceNotFoundException("Could not find any Recipe with id [%s]".formatted(id));
+        }
         boolean hasChanges = false;
 
         if (recipeUpdateRequest.name() != null && !recipeUpdateRequest.name().equals(recipe.getName())) {
@@ -119,23 +120,21 @@ public class RecipeService {
 
         if (recipeUpdateRequest.ingredients() != null && !recipeUpdateRequest.ingredients().isEmpty()
         && !recipeUpdateRequest.ingredients().equals(listOriginalIngredients)) {
-            deleteAllIngredients(recipe.getIdrecipe());
-            ingredientRepository.saveAll(convertToIngredientsList(recipeUpdateRequest.ingredients(), recipe));
+            deleteAllIngredientsByRecipeId(recipe.getIdrecipe());
+            ingredientRepository.saveAll(convertToIngredientsList(recipeUpdateRequest.ingredients(),
+                    recipe,
+                    updateDate));
 
             hasChanges = true;
         }
 
-        List<String> listOriginalSteps = recipe.getSteps()
-                        .stream()
-                        .map(s -> s.getDescription())
-                        .collect(Collectors.toList());
+        String originalStepDescription = recipe.getSteps().getDescription();
 
         if (recipeUpdateRequest.steps() != null && !recipeUpdateRequest.steps().isEmpty()
-        && !recipeUpdateRequest.steps().equals(listOriginalSteps)) {
-
-            deleteAllSteps(recipe.getIdrecipe());
-            stepsRepository.saveAll(convertToStepsList(recipeUpdateRequest.steps(), recipe));
-
+        && !recipeUpdateRequest.steps().equals(originalStepDescription)) {
+            Steps updatedSteps = recipe.getSteps();
+            updatedSteps.setDescription(recipeUpdateRequest.steps());
+            stepsRepository.save(updatedSteps);
             hasChanges = true;
         }
 
@@ -144,59 +143,47 @@ public class RecipeService {
         if (!hasChanges) {
             throw new RequestValidationException("No changes found in recipe with id [%s]".formatted(id));
         }
-        recipe.setUpdatedate(new Date());
+        recipe.setUpdatedate(updateDate);
         recipeRepository.save(recipe);
     }
 
     public void deleteRecipe(Long id) {
-        getRecipeById(id)
-                .orElseThrow(
-                        () -> new ResourceNotFoundException("Could not find any Recipe with id [%s]".formatted(id))
-                );
-        deleteAllIngredients(id);
-
-        deleteAllSteps(id);
+        if (!recipeRepository.existsRecipeByIdrecipe(id)) {
+            throw new ResourceNotFoundException("Could not find any Recipe with id [%s]".formatted(id));
+        }
+        deleteAllIngredientsByRecipeId(id);
+        deleteAllStepsByRecipeId(id);
         recipeRepository.deleteById(id);
     }
 
-    private List<Ingredients> convertToIngredientsList(List<String> inputList, Recipe recipe) {
+    private List<Ingredients> convertToIngredientsList(List<String> inputList, Recipe recipe, Date createDate) {
         List<Ingredients> ingredientsList = new ArrayList<>();
 
         for (String i : inputList) {
-            Ingredients ingredients = new Ingredients();
-            ingredients.setName(i.toLowerCase());
-            ingredients.setCreatedate(new Date());
-            ingredients.setRecipe(recipe);
+            Ingredients ingredients = new Ingredients(
+                    i.toLowerCase(),
+                    createDate,
+                    recipe
+            );
             ingredientsList.add(ingredients);
         }
         return ingredientsList;
     }
 
-    private List<Steps> convertToStepsList(List<String> inputList, Recipe recipe) {
-        List<Steps> stepsList = new ArrayList<>();
-
-        for (String s : inputList) {
-            Steps steps = new Steps();
-            steps.setDescription(s.toLowerCase());
-            steps.setCreatedate(new Date());
-            steps.setRecipe(recipe);
-            stepsList.add(steps);
-        }
-        return stepsList;
-    }
-
-    private void deleteAllIngredients(Long id) {
+    private void deleteAllIngredientsByRecipeId(Long id) {
         List<Ingredients> ingredientsList = ingredientRepository.findAllIngredientsByRecipeId(id);
 
-        if (!ingredientsList.isEmpty())
+        if (!ingredientsList.isEmpty()) {
             ingredientRepository.deleteAll(ingredientsList);
+        }
     }
 
-    private void deleteAllSteps(Long id) {
+    private void deleteAllStepsByRecipeId(Long id) {
         List<Steps> stepsList = stepsRepository.findAllStepsByRecipeId(id);
 
-        if (!stepsList.isEmpty())
+        if (!stepsList.isEmpty()) {
             stepsRepository.deleteAll(stepsList);
+        }
     }
 
 }
